+++ CHANGELOG.md
@@ -204,6 +204,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - `DYNSCREEN_LOCAL_TITLE`
     - `DYNSCREEN_LOCAL_TEXT`
     - `mxdynscreenlocal`
+- Module blacklist-filter functionality
+    - `OxidEsales\EshopCommunity\Internal\Framework\Module\Install\DataObject\OxidEshopPackage::getBlackListFilters()`
+    - `OxidEsales\EshopCommunity\Internal\Framework\Module\Install\DataObject\OxidEshopPackage::setBlackListFilters()`
+    
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)